refactor algorithm task into seperate fn
